+++ vision/deskew/coordinate_lengths.py
